+++ src/utils.ts
@@ -92,15 +92,11 @@ export const getScriptValue = (metric: MetricAggregationWithInlineScript) =>
   (typeof metric.settings?.script === 'object' ? metric.settings?.script?.inline : metric.settings?.script) || '';
 
 export const isSupportedVersion = (version: SemVer): boolean => {
-  if (gte(version, '7.16.0')) {
-    return true;
-  }
-
-  return false;
+  return gte(version, '8.0.0');
 };
 
 export const unsupportedVersionMessage =
-  'Support for Elasticsearch versions after their end-of-life (currently versions < 7.16) was removed. Using unsupported version of Elasticsearch may lead to unexpected and incorrect results.';
+  'Support for Elasticsearch versions after their end-of-life (currently versions < 8.0) was removed. Using unsupported version of Elasticsearch may lead to unexpected and incorrect results.';
 
 // To be considered a time series query, the last bucked aggregation must be a Date Histogram
 export const isTimeSeriesQuery = (query: ElasticsearchDataQuery): boolean => {